Now showing the user's command before showing the response
Because of how requests are handled in AppEngine, you can't just kick
off a goroutine to send the delayed response after responding to allow
the user's original command to show. Instead, this is now using the
appengine/delay package to queue up a function call after the request
is handled. This uses the AppEngine task API under the hood, posting
back to different URL managed by the task system.